requirements.txt
tensorflow>=2.0
numpy
pillow
scikit-learn
matplotlib